Create Submitting Batch Job to Supercomputer
This is the basic script that you will use to run large jobs on the super computer. There are a few things that need to be changed however.
Line 14:  please either replace the email with your own (if you want to receive email notices of jobs that have failed) or delete lines 14-15. You can also set it so that you can receive emails when jobs start and finish, as described on the supercomputer website.

17:  Be sure to change the path to where your version of freesurfer is located, and which reflects which version you are using (if desired)

19:  Be sure to change the path so that it reflects your own SUBJECTS_DIR

21:  You can modify this line so that it executes that command or part of the processing stream that you desire.  The '%1$s' and '%2$s' signify the 1st and 2nd fields in your control file (see the example as shown in 'input_Gale_6-26-13.csv', where %1$s represents the pathway down to your dicoms, and %2$s represents the name you want this subject to have after it has been processed.

37:  If you modify this script for running just part of recon-all (as shown in 'script_gen.submit_v5.1.0_autorecon2-wm.pl'), you will need to modify this line to reflect the number of fields in your control file.  Otherwise the processing won't start and you'll get an error message stating that there is the wrong number of fields in the input line.


Then to set off the script, just type (for example):

./script_gen.submit_v5.1.0.pl < input_Gale_6-26-13.csv